backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
backoff==2.2.1
bidict==0.23.1
blinker==1.8.2
branca==0.8.1
certifi==2024.8.30
cffi==1.16.0
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Django==5.0.3
django-allauth==65.5.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
eventlet==0.38.0
Flask==3.0.3
Flask-SocketIO==5.4.1
flatbuffers==24.3.25
folium==0.19.4
fonttools==4.53.1
frozenlist==1.5.0
gast==0.6.0
google-pasta==0.2.0
googlemaps==4.10.0
greenlet==3.1.1
grpcio==1.68.0
h11==0.14.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
joblib==1.4.2
keras==3.6.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.4.0
MouseInfo==0.1.3
msgpack==1.1.0
multidict==6.1.0
namex==0.0.8
numpy==2.0.0
opencage==3.0.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.13.1
packaging==24.1
phonenumbers==8.13.54
pillow==10.4.0
propcache==0.2.1
protobuf==4.25.3
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.9.0
PyMsgBox==1.0.9
pyobjc-core==10.2 ; sys_platform == 'darwin'
pyobjc-framework-Cocoa==10.2 ; sys_platform == 'darwin'
pyobjc-framework-Quartz==10.2 ; sys_platform == 'darwin'
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-socketio==5.11.4
pytweening==1.2.0
redis==5.2.1
requests==2.32.3
rich==13.9.4
rubicon-objc==0.4.7
scikit-learn==1.5.1
scipy==1.14.0
setuptools==75.6.0
simple-websocket==1.1.0
six==1.16.0
sounddevice==0.4.7
sqlparse==0.4.4
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.67.1
twilio==9.5.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.45.1
wrapt==1.17.0
wsproto==1.2.0
xyzservices==2025.1.0
yarl==1.18.3
geopy
daphne
python-dotenv
psycopg2-binary>=2.9.9
